+++ core/src/config/Networks.hpp
@@ -33,6 +33,7 @@
 
 #include <api/BitcoinLikeNetworkParameters.hpp>
 #include <api/EthereumLikeNetworkParameters.hpp>
+#include <api/RippleLikeNetworkParameters.hpp>
 #include <api/Networks.hpp>