+++ core/src/jni/jni/Operation.cpp
@@ -9,6 +9,7 @@
 #include "Marshal.hpp"
 #include "OperationType.hpp"
 #include "Preferences.hpp"
+#include "RippleLikeOperation.hpp"
 #include "TrustIndicator.hpp"
 #include "WalletType.hpp"
 
@@ -154,6 +155,16 @@ CJNIEXPORT jobject JNICALL Java_co_ledger_core_Operation_00024CppProxy_native_1a
         const auto& ref = ::djinni::objectFromHandleAddress<::ledger::core::api::Operation>(nativeRef);
         auto r = ref->asEthereumLikeOperation();
         return ::djinni::release(::djinni_generated::EthereumLikeOperation::fromCpp(jniEnv, r));
+    } JNI_TRANSLATE_EXCEPTIONS_RETURN(jniEnv, 0 /* value doesn't matter */)
+}
+
+CJNIEXPORT jobject JNICALL Java_co_ledger_core_Operation_00024CppProxy_native_1asRippleLikeOperation(JNIEnv* jniEnv, jobject /*this*/, jlong nativeRef)
+{
+    try {
+        DJINNI_FUNCTION_PROLOGUE1(jniEnv, nativeRef);
+        const auto& ref = ::djinni::objectFromHandleAddress<::ledger::core::api::Operation>(nativeRef);
+        auto r = ref->asRippleLikeOperation();
+        return ::djinni::release(::djinni_generated::RippleLikeOperation::fromCpp(jniEnv, r));
     } JNI_TRANSLATE_EXCEPTIONS_RETURN(jniEnv, 0 /* value doesn't matter */)
 }