+++ core/src/utils/DateUtils.hpp
@@ -39,6 +39,7 @@ namespace ledger {
         class DateUtils {
         public:
             static std::chrono::system_clock::time_point fromJSON(const std::string& str);
+            static std::string formatDateFromJSON(const std::string& str);
             static std::string toJSON(const std::chrono::system_clock::time_point& date);
             static std::chrono::system_clock::time_point now();
             static std::chrono::system_clock::time_point incrementDate(const std::chrono::system_clock::time_point &date, api::TimePeriod precision);