+++ core/src/wallet/ethereum/factories/EthereumLikeKeychainFactory.h
@@ -50,13 +50,6 @@ namespace ledger {
                                                         const api::ExtendedKeyAccountCreationInfo& info,
                                                         const std::shared_ptr<Preferences>& accountPreferences,
                                                         const api::Currency& currency);
-
-//            std::shared_ptr<EthereumLikeKeychain> restore(int32_t index,
-//                                                         const DerivationPath &path,
-//                                                         const std::shared_ptr<DynamicObject>& configuration,
-//                                                         const std::string& address,
-//                                                         const std::shared_ptr<Preferences>& accountPreferences,
-//                                                         const api::Currency& currency);
             std::shared_ptr<EthereumLikeKeychain> restore(int32_t index,
                                                           const DerivationPath &path,
                                                           const std::shared_ptr<DynamicObject>& configuration,